+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/triples/impl/BitmapTriplesIteratorZ.java
@@ -46,7 +46,7 @@ public class BitmapTriplesIteratorZ implements IteratorTripleID {
 	
 	private AdjacencyList adjY, adjZ;
 	long posZ;
-	private int x, y, z;
+	private long x, y, z;
 	
 	BitmapTriplesIteratorZ(BitmapTriples triples, TripleID pattern) {
 		this.triples = triples;
@@ -87,9 +87,9 @@ public class BitmapTriplesIteratorZ implements IteratorTripleID {
 		long posY = adjZ.findListIndex(posZ);
 		
 		// Set full triple
-		z = (int) adjZ.get(posZ);
-		y = (int) adjY.get(posY);
-		x = (int) adjY.findListIndex(posY)+1;
+		z = adjZ.get(posZ);
+		y = adjY.get(posY);
+		x = adjY.findListIndex(posY)+1;
 		
 		// Go forward finding next appearance
 		posZ = adjZ.findNextAppearance(posZ+1, patZ);
